+++ packages/paste-core/components/card/__test__/card.test.tsx
@@ -7,7 +7,7 @@ describe('Card', () => {
   it('it should render default values', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Card />
         </Theme.Provider>
       )
@@ -15,20 +15,21 @@ describe('Card', () => {
     expect(tree).toMatchSnapshot();
   });
 
-  it('it should error with bad propTypes', (): void => {
-    expect(() => {
-      renderer.create(
-        <Theme.Provider>
-          <Card error="blowitup" />
+  it('it should filter out style props that are not allowed', (): void => {
+    const tree = renderer
+      .create(
+        <Theme.Provider theme="console">
+          <Card margin="space10" backgroundColor="colorBackgroundSuccess" />
         </Theme.Provider>
-      );
-    }).toThrow();
+      )
+      .toJSON();
+    expect(tree).toMatchSnapshot();
   });
 
   it('it should render default values unless overridden by the component', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Card paddingBottom="space200" />
         </Theme.Provider>
       )
@@ -38,7 +39,7 @@ describe('Card', () => {
   it('it should render children', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Card>I AM A JEDI!!!!</Card>
         </Theme.Provider>
       )