+++ packages/paste-core/components/paragraph/__tests__/paragraph.test.tsx
@@ -7,7 +7,7 @@ describe('Paragraph', () => {
   it('it should render a single paragraph', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Paragraph>This is a paragraph</Paragraph>
         </Theme.Provider>
       )
@@ -18,7 +18,7 @@ describe('Paragraph', () => {
   it('it should render a single paragraph with italic text', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Paragraph>
             <i>This is a paragraph with italic text</i>
           </Paragraph>
@@ -31,7 +31,7 @@ describe('Paragraph', () => {
   it('it should render a single paragraph with bold text', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Paragraph>
             <strong>This is a paragraph with bold text</strong>
           </Paragraph>