+++ packages/paste-core/components/spinner/stories/index.stories.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import {storiesOf} from '@storybook/react';
 import {withKnobs, text, select, boolean} from '@storybook/addon-knobs';
-import {TextColor, IconSize} from '@twilio-paste/types';
+import {IconSize, TextColor} from '@twilio-paste/style-props';
 import {DefaultTheme} from '@twilio-paste/theme-tokens';
 import {Spinner} from '../src';