+++ packages/paste-core/utilities/text/__tests__/SafelySpreadProps.spec.ts
@@ -1,0 +1,26 @@
+import {safelySpreadTextProps} from '../src/SafelySpreadProps';
+
+const badProps = {
+  foo: 'bar',
+  textColor: 'textColorSuccess',
+};
+
+const goodProps = {
+  baz: 'foo',
+  'aria-label': 'test',
+};
+
+describe('safelySpreadTextProps', () => {
+  it('should not filter out good props', () => {
+    expect(safelySpreadTextProps(goodProps)).toEqual({
+      baz: 'foo',
+      'aria-label': 'test',
+    });
+  });
+
+  it('should filter out bad props', () => {
+    expect(safelySpreadTextProps(badProps)).toEqual({
+      foo: 'bar',
+    });
+  });
+});