+++ packages/paste-core/utilities/text/__tests__/text.spec.tsx
@@ -7,7 +7,7 @@ describe('as', () => {
   it('it should render as a provided HTML element', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="label">background single</Text>
         </Theme.Provider>
       )
@@ -20,7 +20,7 @@ describe('display', () => {
   it('should set a display property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" display="inline-block">
             display single
           </Text>
@@ -33,7 +33,7 @@ describe('display', () => {
   it('should set a responsive display property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" display={['inline-block', 'block']}>
             display responsive
           </Text>
@@ -48,7 +48,7 @@ describe('fontFamily', () => {
   it('should set a font family property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" fontFamily="fontFamilyCode">
             font family single
           </Text>
@@ -61,7 +61,7 @@ describe('fontFamily', () => {
   it('should set a responsive font family property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" fontFamily={['fontFamilyText', 'fontFamilyCode']}>
             font family responsive
           </Text>
@@ -76,7 +76,7 @@ describe('fontSize', () => {
   it('should set a font size property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" fontSize="fontSize10">
             font size single
           </Text>
@@ -89,7 +89,7 @@ describe('fontSize', () => {
   it('should set a responsive font size property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" fontSize={['fontSize20', 'fontSize30']}>
             font size responsive
           </Text>
@@ -104,7 +104,7 @@ describe('fontStyle', () => {
   it('should set a font style property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" fontStyle="italic">
             font style single
           </Text>
@@ -117,7 +117,7 @@ describe('fontStyle', () => {
   it('should set a responsive font style property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" fontStyle={['italic', 'normal']}>
             font style responsive
           </Text>
@@ -132,7 +132,7 @@ describe('fontWeight', () => {
   it('should set a font weight property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" fontWeight="fontWeightBold">
             font weight single
           </Text>
@@ -145,7 +145,7 @@ describe('fontWeight', () => {
   it('should set a responsive font weight property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" fontWeight={['fontWeightLight', 'fontWeightSemibold']}>
             font weight responsive
           </Text>
@@ -160,7 +160,7 @@ describe('letterSpacing', () => {
   it('should set a letter spacing property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" letterSpacing="1px">
             letter spacing single
           </Text>
@@ -173,7 +173,7 @@ describe('letterSpacing', () => {
   it('should set a responsive letter spacing property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" letterSpacing={['-2px', '3px']}>
             letter spacing responsive
           </Text>
@@ -188,7 +188,7 @@ describe('lineHeight', () => {
   it('should set a line height property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" lineHeight="lineHeight10">
             line height single
           </Text>
@@ -201,7 +201,7 @@ describe('lineHeight', () => {
   it('should set a responsive line height property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" lineHeight={['lineHeight20', 'lineHeight30']}>
             line height responsive
           </Text>
@@ -216,7 +216,7 @@ describe('margin', () => {
   it('should set a margin property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" margin="space10">
             margin single
           </Text>
@@ -229,7 +229,7 @@ describe('margin', () => {
   it('should set a responsive margin property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" margin={['space20', 'space30']}>
             margin responsive
           </Text>
@@ -244,7 +244,7 @@ describe('marginBottom', () => {
   it('should set a marginBottom property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" marginBottom="space10">
             marginBottom single
           </Text>
@@ -257,7 +257,7 @@ describe('marginBottom', () => {
   it('should set a responsive marginBottom property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" marginBottom={['space20', 'space30']}>
             marginBottom responsive
           </Text>
@@ -272,7 +272,7 @@ describe('marginLeft', () => {
   it('should set a marginLeft property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" marginLeft="space10">
             marginLeft single
           </Text>
@@ -285,7 +285,7 @@ describe('marginLeft', () => {
   it('should set a responsive marginLeft property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" marginLeft={['space20', 'space30']}>
             marginLeft responsive
           </Text>
@@ -300,7 +300,7 @@ describe('marginRight', () => {
   it('should set a marginRight property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" marginRight="space10">
             marginRight single
           </Text>
@@ -313,7 +313,7 @@ describe('marginRight', () => {
   it('should set a responsive marginRight property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" marginRight={['space20', 'space30']}>
             marginRight responsive
           </Text>
@@ -328,7 +328,7 @@ describe('marginTop', () => {
   it('should set a marginTop property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" marginTop="space10">
             marginTop single
           </Text>
@@ -341,7 +341,7 @@ describe('marginTop', () => {
   it('should set a responsive marginTop property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" marginTop={['space20', 'space30']}>
             marginTop responsive
           </Text>
@@ -356,7 +356,7 @@ describe('padding', () => {
   it('should set a padding property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" padding="space10">
             padding single
           </Text>
@@ -369,7 +369,7 @@ describe('padding', () => {
   it('should set a responsive padding property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" padding={['space20', 'space30']}>
             padding responsive
           </Text>
@@ -384,7 +384,7 @@ describe('paddingBottom', () => {
   it('should set a paddingBottom property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" paddingBottom="space10">
             paddingBottom single
           </Text>
@@ -397,7 +397,7 @@ describe('paddingBottom', () => {
   it('should set a responsive paddingBottom property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" paddingBottom={['space20', 'space30']}>
             paddingBottom responsive
           </Text>
@@ -412,7 +412,7 @@ describe('paddingLeft', () => {
   it('should set a paddingLeft property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" paddingLeft="space10">
             paddingLeft single
           </Text>
@@ -425,7 +425,7 @@ describe('paddingLeft', () => {
   it('should set a responsive paddingLeft property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" paddingLeft={['space20', 'space30']}>
             paddingLeft responsive
           </Text>
@@ -440,7 +440,7 @@ describe('paddingRight', () => {
   it('should set a paddingRight property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" paddingRight="space10">
             paddingRight single
           </Text>
@@ -453,7 +453,7 @@ describe('paddingRight', () => {
   it('should set a responsive paddingRight property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" paddingRight={['space20', 'space30']}>
             paddingRight responsive
           </Text>
@@ -468,7 +468,7 @@ describe('paddingTop', () => {
   it('should set a paddingTop property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" paddingTop="space10">
             paddingTop single
           </Text>
@@ -481,7 +481,7 @@ describe('paddingTop', () => {
   it('should set a responsive paddingTop property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" paddingTop={['space20', 'space30']}>
             paddingTop responsive
           </Text>
@@ -496,7 +496,7 @@ describe('textAlign', () => {
   it('should set a textAlign property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" textAlign="right">
             textAlign single
           </Text>
@@ -509,7 +509,7 @@ describe('textAlign', () => {
   it('should set a responsive textAlign property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" textAlign={['center', 'left']}>
             textAlign responsive
           </Text>
@@ -524,7 +524,7 @@ describe('textColor', () => {
   it('should set a textColor property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" textColor="colorText">
             textColor single
           </Text>
@@ -537,7 +537,7 @@ describe('textColor', () => {
   it('should set a responsive textColor property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" textColor={['colorTextError', 'colorTextSuccess']}>
             textColor responsive
           </Text>
@@ -552,7 +552,7 @@ describe('textDecoration', () => {
   it('should set a textDecoration property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" textDecoration="underline">
             textDecoration single
           </Text>
@@ -565,7 +565,7 @@ describe('textDecoration', () => {
   it('should set a responsive textDecoration property', (): void => {
     const tree = renderer
       .create(
-        <Theme.Provider>
+        <Theme.Provider theme="console">
           <Text as="span" textDecoration={['underline', 'normal']}>
             textDecoration responsive
           </Text>