+++ packages/paste-icons/rollup.icon-list.js
@@ -1,1 +1,1 @@
-export const rollupIconList = ['src/LoadingIcon.tsx', 'src/PlusIcon.tsx'];
+export const rollupIconList = ["src/CloseIcon.tsx","src/CopyIcon.tsx","src/ErrorIcon.tsx","src/InformationIcon.tsx","src/LoadingIcon.tsx","src/PlusIcon.tsx","src/WarningIcon.tsx"];