+++ packages/paste-core/components/list/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.1.0](https://github.com/twilio-labs/paste/compare/@twilio-paste/list@0.0.2...@twilio-paste/list@0.1.0) (2019-12-19)
+
+
+### Features
+
+* **list:** created the list component ([bd6db2e](https://github.com/twilio-labs/paste/commit/bd6db2e7de6a72fda8d42455f12623ca51440e27))
+
+
+
+
+
 ## [0.0.2](https://github.com/twilio-labs/paste/compare/@twilio-paste/list@0.0.1...@twilio-paste/list@0.0.2) (2019-10-29)
 
 **Note:** Version bump only for package @twilio-paste/list