+++ packages/paste-theme/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.5](https://github.com/twilio-labs/paste/compare/@twilio-paste/theme@2.0.4...@twilio-paste/theme@2.0.5) (2019-12-19)
+
+**Note:** Version bump only for package @twilio-paste/theme
+
+
+
+
+
 ## [2.0.4](https://github.com/twilio-labs/paste/compare/@twilio-paste/theme@2.0.3...@twilio-paste/theme@2.0.4) (2019-12-05)
 
 **Note:** Version bump only for package @twilio-paste/theme